Move .defer tests to separate file.
